FIX |#24| Move api_responses.py to the resources folder
